+++ Modules/Project.Windows.ProgramInfo/Public/Test-UserProfile.ps1
@@ -30,8 +30,7 @@ SOFTWARE.
 .SYNOPSIS
 Check if input path leads to user profile
 .DESCRIPTION
-User profile paths are not valid for firewall rules, this method help make a check
-if this is true
+User profile paths are not valid for firewall rules, this method help to check if this is true
 .PARAMETER FilePath
 File path to check, can be unformatted or have environment variables
 .EXAMPLE
@@ -42,6 +41,7 @@ None. You cannot pipe objects to Test-UserProfile
 [bool] true if userprofile path or false otherwise
 .NOTES
 TODO: is it possible to nest this into Test-Environment somehow?
+TODO: This should proably be inside Utility or UserInfo module
 #>
 function Test-UserProfile
 {