+++ Templates/ScriptTemplate.ps1
@@ -66,7 +66,8 @@ the Functionality parameter of Get-Help.
 	"PSReviewUnusedParameter", "", Justification = "This is template script parameter")]
 [CmdletBinding()]
 param (
-	$ScriptParam
+	[Parameter()]
+	[ParameterType] $ScriptParam
 )
 
 # Initialization
@@ -78,6 +79,7 @@ New-Variable -Name ThisScript -Scope Private -Option Constant -Value (
 
 # Check requirements
 Initialize-Project -Abort
+Write-Debug -Message "[$ThisScript] params($($PSBoundParameters.Values))"
 
 # Imports
 # TODO: Include modules and scripts as needed